Revert "PM-439 Move actions, reducers related to provider integrations to src/integrations; Start working on using immutable implementation"
This reverts commit 9e11b62b02cbe568a4033804615619d45097d31e.
