--- src/integrations/store/models/index.js
@@ -1,1 +1,0 @@
-export { default as ProviderRecord } from './provider'
