--- src/integrations/store/models/provider.js
@@ -1,13 +1,0 @@
-import { Record } from 'immutable'
-
-const ProviderRecord = Record({
-  name: undefined,
-  available: false,
-  loaded: false,
-  network: undefined,
-  account: null,
-  priority: undefined,
-  networkId: 0,
-}, 'Provider')
-
-export default ProviderRecord
