--- src/integrations/store/reducers/providers.js
@@ -1,5 +1,0 @@
-import { Map } from 'immutable'
-import { handleActions } from 'redux-actions'
-import { ProviderRecord } from '../models'
-
-export default handleActions({}, Map())
