+++ src/selectors/event.js
@@ -1,5 +1,5 @@
 import { get } from 'lodash'
-import { entitySelector } from 'selectors/entities'
+import { entitySelector } from './entities'
 
 export const EVENT_TYPE_CATEGORICAL = 'CATEGORICAL'
 export const EVENT_TYPE_SCALAR = 'SCALAR'
