+++ src/selectors/marketShares.js
@@ -5,11 +5,11 @@ import moment from 'moment'
 import { add0xPrefix, calcShareWinnings } from 'utils/helpers'
 import { calcLMSRMarginalPrice, calcLMSROutcomeTokenCount } from 'api'
 
-import { getCurrentAccount } from 'integrations/store/selectors'
-import { getEvents } from 'selectors/event'
-import { getOracles } from 'selectors/oracle'
-import { getEventDescriptions } from 'selectors/eventDescription'
-import { MARKET_STAGES } from 'utils/constants'
+import { getCurrentAccount } from 'integrations/selectors'
+import { getEvents } from './event'
+import { getOracles } from './oracle'
+import { getEventDescriptions } from './eventDescription'
+import { MARKET_STAGES } from '../utils/constants'
 
 export const getShares = (state) => {
   if (!state.entities) {
